requirements.txt
uvloop
uvicorn
fastapi
requests
kurigram
tgcrypto
python-dotenv